backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI & LLM
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.50.0

# Vector Store
chromadb==0.4.22

# Ollama Support
ollama==0.1.6

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
pandas==2.2.0
openpyxl==3.1.2

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
scrubadub==2.0.1

# Development
pytest==8.0.0
pytest-asyncio==0.23.3
black==24.1.1
